+++ scripts/plugin-tests.js
@@ -2305,6 +2305,93 @@ let syncTests = {
 
     result.rebuild.dispose()
   },
+
+  async onStartOnEndWatchMode({ esbuild, testDir }) {
+    const srcDir = path.join(testDir, 'src')
+    const outfile = path.join(testDir, 'out.js')
+    const input = path.join(srcDir, 'in.js')
+    const example = path.join(srcDir, 'example.js')
+    await mkdirAsync(srcDir, { recursive: true })
+    await writeFileAsync(input, `import {x} from "./example.js"; exports.x = x`)
+    await writeFileAsync(example, `export let x = 1`)
+
+    let onStartCalls = 0
+    let onEndCalls = 0
+    let onRebuild = () => { }
+
+    const result = await esbuild.build({
+      entryPoints: [input],
+      outfile,
+      format: 'cjs',
+      logLevel: 'silent',
+      watch: {
+        onRebuild: (...args) => onRebuild(args),
+      },
+      bundle: true,
+      plugins: [
+        {
+          name: 'some-plugin',
+          setup(build) {
+            build.onStart(() => {
+              onStartCalls++
+            })
+            build.onEnd(() => {
+              onEndCalls++
+            })
+
+            build.onLoad({ filter: /example\.js$/ }, async (args) => {
+              const contents = await fs.promises.readFile(args.path, 'utf8')
+              return { contents }
+            })
+          },
+        },
+      ],
+    })
+
+    assert.strictEqual(onStartCalls, 1)
+    assert.strictEqual(onEndCalls, 1)
+
+    const rebuildUntil = (mutator, condition) => {
+      let timeout
+      return new Promise((resolve, reject) => {
+        timeout = setTimeout(() => reject(new Error('Timeout after 30 seconds')), 30 * 1000)
+        onRebuild = args => {
+          try { if (condition(...args)) clearTimeout(timeout), resolve(args) }
+          catch (e) { clearTimeout(timeout), reject(e) }
+        }
+        mutator()
+      })
+    }
+
+    try {
+      let code = await readFileAsync(outfile, 'utf8')
+      let exports = {}
+      new Function('exports', code)(exports)
+      assert.strictEqual(result.outputFiles, void 0)
+      assert.strictEqual(typeof result.stop, 'function')
+      assert.strictEqual(exports.x, 1)
+
+      // First rebuild: edit
+      {
+        const [error2, result2] = await rebuildUntil(
+          () => setTimeout(() => writeFileAtomic(example, `export let x = 2`), 250),
+          () => fs.readFileSync(outfile, 'utf8') !== code,
+        )
+        code = await readFileAsync(outfile, 'utf8')
+        exports = {}
+        new Function('exports', code)(exports)
+        assert.strictEqual(error2, null)
+        assert.strictEqual(result2.outputFiles, void 0)
+        assert.strictEqual(result2.stop, result.stop)
+        assert.strictEqual(exports.x, 2)
+      }
+
+      assert.notStrictEqual(onStartCalls, 1)
+      assert.notStrictEqual(onEndCalls, 1)
+    } finally {
+      result.stop()
+    }
+  },
 }
 
 async function main() {